requirements.txt
pandas==3.0.1
openpyxl==3.1.3
python-dotenv==1.2.2
requests==2.32.5
tk